// src/lexic/enum/token-types.enum.ts
export enum TokenTypes {
  WHITE_SPACE = 'ws',
  LINE_BREAK = 'lb',
  COMMENT = 'comment',
  MULTILINE_COMMENT = 'multiline-comment',
  IDENTIFIER = 'identifier',
  KEYWORD = 'keyword',
  PROCEDURE = 'procedure',
  TYPE = 'type',
  DOUBLE = 'double',
  INTEGER = 'integer',
  STRING = 'string',
  PLUS = 'plus',
  MINUS = 'minus',
  TIMES = 'times',
  DIVIDE = 'divide',
  ASSIGN = 'assign',
  EQUALS = 'equals',
  NOT_EQUALS = 'not-equals',
  LESS_THAN = 'less-than',
  LESS_THAN_EQUALS = 'less-than-equals',
  GREATER_THAN = 'greater-than',
  GREATER_THAN_EQUALS = 'greater-than-equals',
  L_PAREN = 'l-paren',
  R_PAREN = 'r-paren',
  SEMICOLON = 'semicolon',
  COLON = 'colon',
  COMMA = 'comma',
  DOT = 'dot',
  ERROR = 'error',
}